· Added empty readers ( DictionaryReader, EffectMaterialReader, ExternalReferenceReader, TextureReader ) · Added empty effect material class · Changes to ContentResourceManager · Added IDisposable to Model class

+++ SceneR.Framework/Sources/Framework/RendererWindow.cpp
@@ -79,7 +79,7 @@ void RendererWindow::Open()
     glfwWindowHint(GLFW_ALPHA_BITS            , 8);
     glfwWindowHint(GLFW_DEPTH_BITS            , 24);
     glfwWindowHint(GLFW_STENCIL_BITS          , 24);
-    glfwWindowHint(GLFW_SAMPLES               , sampleCount);
+    glfwWindowHint(GLFW_SAMPLES               , static_cast<Int32>(sampleCount));
     glfwWindowHint(GLFW_SRGB_CAPABLE          , true);
 
     if (fullscreen)
@@ -90,11 +90,11 @@ void RendererWindow::Open()
     // Create a new window
     this->handle = glfwCreateWindow
     (
-        width       // width
-      , height      // height
-      , tmp.c_str() // title
-      , monitor     // monitor
-      , windowShare // share
+        static_cast<Int32>(width)   // width
+      , static_cast<Int32>(height)  // height
+      , tmp.c_str()                 // title
+      , monitor                     // monitor
+      , windowShare                 // share
     );
 
     // If glfwCreateWindow is failing for you, then you may need to lower the OpenGL version.
